Generator for the function set of the real-parameter Bi-Objective Black-Box Optimization Benchmarking (BBOB).

Generator for the function set of the real-parameter Bi-Objective Black-Box Optimization Benchmarking (BBOB).

makeBiObjBBOBFunction(dimensions, fid, iid)

Arguments

dimensions
[integer(1)]
Problem dimensions. Integer value between 2 and 40.

fid
[integer(1)]
Function identifier. Integer value between 1 and 55.

iid
[integer(1)]
Instance identifier. Integer value greater than or equal 1.

Value

[smoof_multi_objective_function]

Note

Concatenation of single-objective BBOB functions into a bi-objective problem.

References

See the COCO website for a detailed description of the bi-objective BBOB functions. An overview of which pair of single-objective BBOB functions creates which of the 55 bi-objective BBOB functions can be found here.

Examples

# get the fifth instance of the concatenation of the
# 3D versions of sphere and Rosenbrock
fn = makeBiObjBBOBFunction(dimensions = 3L, fid = 4L, iid = 5L)
fn(c(3, -1, 0))
#> [1]    229.5489 102360.1058
# compare to the output of its single-objective pendants
f1 = makeBBOBFunction(dimensions = 3L, fid = 1L, iid = 2L * 5L + 1L)
f2 = makeBBOBFunction(dimensions = 3L, fid = 8L, iid = 2L * 5L + 2L)
identical(fn(c(3, -1, 0)), c(f1(c(3, -1, 0)), f2(c(3, -1, 0))))
#> [1] TRUE