ED1 Function

Builds and returns the multi-objective ED1 test problem.

The ED1 test problem is defined as follows:

Minimize \(f_j(\mathbf{x}) = \frac{1}{r(\mathbf{x}) + 1} \cdot \tilde{p}(\Theta (\mathbf{X}))\), for \(j = 1, \ldots, m\),

with \(\mathbf{x} = (x_1, \ldots, x_n)^T\), where \(0 \leq x_i \leq 1\), and \(\Theta = (\theta_1, \ldots, \theta_{m-1})\), where \(0 \le \theta_j \le \frac{\pi}{2}\), for \(i = 1, \ldots, n,\) and \(j = 1, \ldots, m - 1\).

Moreover \(r(\mathbf{X}) = \sqrt{x_m^2 + \ldots, x_n^2}\),

\(\tilde{p}_1(\Theta) = \cos(\theta_1)^{2/\gamma}\),

\(\tilde{p}_j(\Theta) = \left( \sin(\theta_1) \cdot \ldots \cdot \sin(\theta_{j - 1}) \cdot \cos(\theta_j) \right)^{2/\gamma}\), for \(2 \le j \le m - 1\),

and \(\tilde{p}_m(\Theta) = \left( \sin(\theta_1) \cdot \ldots \cdot \sin(\theta_{m - 1}) \right)^{2/\gamma}\).

makeED1Function(dimensions, n.objectives, gamma = 2, theta)

Arguments

dimensions

[integer(1)]
Number of decision variables.

n.objectives

[integer(1)]
Number of objectives.

gamma

[numeric(1)]
Optional parameter. Default is 2, which is recommended by Emmerich and Deutz.

theta

[numeric(dimensions)]
Parameter vector, whose components have to be between 0 and 0.5*pi. The default is theta = (pi/2) * x (with x being the point from the decision space) as recommended by Emmerich and Deutz.

Value

[smoof_multi_objective_function]

References

M. T. M. Emmerich and A. H. Deutz. Test Problems based on Lame Superspheres. Proceedings of the International Conference on Evolutionary Multi-Criterion Optimization (EMO 2007), pp. 922-936, Springer, 2007.