Generator for function with multiple peaks following the multiple peaks model 2.

Generator for function with multiple peaks following the multiple peaks model 2.

makeMPM2Function(n.peaks, dimensions, topology, seed, rotated = TRUE,
  peak.shape = "ellipse")

Arguments

n.peaks
[integer(1)] Desired number of peaks, i. e., number of (local) optima.

dimensions
[integer(1)] Size of corresponding parameter space.

topology
[character(1)] Type of topology. Possible values are “random” and “funnel”.

seed
[integer(1)] Seed for the random numbers generator.

rotated
[logical(1)] Should the peak shapes be rotated? This parameter is only relevant in case of elliptically shaped peaks.

peak.shape
[character(1)] Shape of peak(s). Possible values are “ellipse” and “sphere”.

Value

[smoof_single_objective_function]

References

See the technical report of multiple peaks model 2 for an in-depth description of the underlying algorithm.

Examples

not_run({
  fn = makeMPM2Function(n.peaks = 10L, dimensions = 2L,
    topology = "funnel", seed = 123, rotated = TRUE, peak.shape = "ellipse")
  if (require(plot3D)) {
    plot3D(fn)
  }
})
not_run({
  fn = makeMPM2Function(n.peaks = 5L, dimensions = 2L,
    topology = "random", seed = 134, rotated = FALSE)
  plot(fn, render.levels = TRUE)
})