WFG9 Function

Ninth test problem from the "Walking Fish Group" problem generator toolkit.

makeWFG9Function(n.objectives, k, l)

Arguments

n.objectives
[integer(1)]
Number of objectives.

k
[integer(1)]
Number of position-related parameters. These will automatically be the first k elements from the input vector. This value has to be a multiple of n.objectives - 1.

l
[integer(1)]
Number of distance-related parameters. These will automatically be the last l elements from the input vector.

Value

[smoof_multi_objective_function]

Details

Huband et al. recommend a value of k = 4L position-related parameters for bi-objective problems and k = 2L * (n.objectives - 1L) for many-objective problems. Furthermore the authors recommend a value of l = 20 distance-related parameters. Therefore, if k and/or l are not explicitly defined by the user, their values will be set to the recommended values per default.

References

S. Huband, P. Hingston, L. Barone, and L. While, "A Review of Multi-objective Test Problems and a Scalable Test Problem Toolkit," in IEEE Transactions on Evolutionary Computation, Volume 10, No 5, October 2006, pp. 477-506. IEEE.