Plot a two-dimensional numeric function.

Either a contour-plot or a level-plot.

plot2DNumeric(
  x,
  show.optimum = FALSE,
  main = getName(x),
  render.levels = FALSE,
  render.contours = TRUE,
  n.samples = 100L,
  ...
)

Arguments

x

[smoof_function]
Function.

show.optimum

[logical(1)]
If the function has a known global optimum, should its location be plotted by a point or multiple points in case of multiple global optima? Default is FALSE.

main

[character(1L)]
Plot title. Default is the name of the smoof function.

render.levels

[logical(1)]
Show a level-plot? Default is FALSE.

render.contours

[logical(1)]
Render contours? Default is TRUE.

n.samples

[integer(1)]
Number of locations per dimension to be sampled. Default is 100.

...

[any]
Further paramerters passed to image respectively contour function.

Value

Nothing