Return a function which internally stores x or y values.

Often it is desired and useful to store the optimization path, i.e., the evaluated function values and/or the parameters. Not all optimization algorithms offer such a trace. This wrapper makes a smoof function handle x/y-values itself.

addLoggingWrapper(fn, logg.x = FALSE, logg.y = TRUE)

Arguments

fn
[smoof_function] Smoof function.

logg.x
[logical(1)] Should x-values be logged? Default is FALSE.

logg.y
[logical(1)] Should objective values be logged? Default is TRUE.

Value

[smoof_logging_function]

Note

Logging values, in particular logging x-values, will substantially slow down the evaluation of the function.

Examples

# We first build the smoof function and apply the logging wrapper to it
fn = makeSphereFunction(dimension = 2L)
fn = addLoggingWrapper(fn, logg.x = TRUE)

# We now apply an optimization algorithm to it and the logging wrapper keeps
# track of the evaluated points.
res = optim(fn, par = c(1, 1), method = "Nelder-Mead")

# Extract the logged values
log.res = getLoggedValues(fn)
print(log.res$pars)

#>            x1         x2
#> 1   1.000e+00  1.000e+00
#> 2   1.100e+00  1.000e+00
#> 3   1.000e+00  1.100e+00
#> 4   9.000e-01  1.100e+00
#> 5   9.500e-01  1.075e+00
#> 6   9.000e-01  1.000e+00
#> 7   8.500e-01  9.500e-01
#> 8   9.500e-01  8.500e-01
#> 9   9.375e-01  9.125e-01
#> 10  8.000e-01  8.000e-01
#> 11  7.000e-01  7.000e-01
#> 12  8.000e-01  6.000e-01
#> 13  8.125e-01  6.875e-01
#> 14  5.500e-01  4.500e-01
#> 15  3.500e-01  2.500e-01
#> 16  2.500e-01  3.500e-01
#> 17  3.875e-01  4.125e-01
#> 18 -1.000e-01 -1.000e-01
#> 19 -5.000e-01 -5.000e-01
#> 20 -2.000e-01 -8.882e-16
#> 21 -6.250e-02  6.250e-02
#> 22 -4.125e-01 -3.875e-01
#> 23  8.437e-02  1.656e-01
#> 24 -2.469e-01 -2.031e-01
#> 25  1.562e-03  7.344e-02
#> 26  3.906e-02  2.359e-01
#> 27 -6.523e-02 -1.602e-02
#> 28 -1.172e-03 -5.078e-03
#> 29  2.949e-02 -3.887e-02
#> 30 -6.797e-02 -9.453e-02
#> 31 -1.582e-02  3.145e-02
#> 32  4.824e-02  4.238e-02
#> 33  1.987e-02  2.778e-02
#> 34  3.452e-02 -8.740e-03
#> 35 -3.235e-03  2.140e-02
#> 36 -2.428e-02 -1.146e-02
#> 37 -1.324e-02 -1.651e-03
#> 38 -1.118e-02 -2.813e-02
#> 39 -5.221e-03  9.017e-03
#> 40  6.849e-03  5.590e-03
#> 41  1.826e-03  3.780e-03
#> 42  5.875e-03 -1.032e-02
#> 43 -2.447e-03  4.184e-03
#> 44  5.514e-04  1.304e-02
#> 45 -7.411e-04 -5.481e-04
#> 46  3.532e-03 -9.523e-04
#> 47  2.037e-03  3.318e-04
#> 48 -5.300e-04 -3.996e-03
#> 49  5.904e-05 -2.052e-03
#> 50 -2.719e-03 -2.932e-03
#> 51  8.482e-04 -4.842e-04
#> 52  4.804e-05  1.020e-03
#> 53  5.079e-05  2.519e-04
#> 54 -1.538e-03  1.879e-04
#> 55  2.515e-04 -3.162e-04
#> 56  1.043e-03  4.838e-04
#> 57 -2.950e-04 -2.901e-04
#> 58  5.973e-04  2.258e-04
#> 59 -7.190e-05 -1.611e-04
#> 60 -2.726e-04  4.069e-04
#> 61  1.205e-04 -1.354e-04
#> 62 -2.220e-06 -5.484e-04
#> 63  3.754e-05  5.179e-05

print(log.res$obj.vals)

#>  [1] 2.000e+00 2.210e+00 2.210e+00 2.020e+00 2.058e+00 1.810e+00 1.625e+00
#>  [8] 1.625e+00 1.712e+00 1.280e+00 9.800e-01 1.000e+00 1.133e+00 5.050e-01
#> [15] 1.850e-01 1.850e-01 3.203e-01 2.000e-02 5.000e-01 4.000e-02 7.813e-03
#> [22] 3.203e-01 3.455e-02 1.022e-01 5.396e-03 5.719e-02 4.512e-03 2.716e-05
#> [29] 2.380e-03 1.356e-02 1.239e-03 4.124e-03 1.167e-03 1.268e-03 4.684e-04
#> [36] 7.209e-04 1.781e-04 9.161e-04 1.086e-04 7.816e-05 1.762e-05 1.409e-04
#> [43] 2.349e-05 1.704e-04 8.496e-07 1.338e-05 4.261e-06 1.625e-05 4.215e-06
#> [50] 1.599e-05 9.538e-07 1.042e-06 6.601e-08 2.402e-06 1.632e-07 1.323e-06
#> [57] 1.712e-07 4.077e-07 3.114e-08 2.399e-07 3.285e-08 3.007e-07 4.092e-09

log.res = getLoggedValues(fn, compact = TRUE)
print(log.res)

#>            x1         x2        y1
#> 1   1.000e+00  1.000e+00 2.000e+00
#> 2   1.100e+00  1.000e+00 2.210e+00
#> 3   1.000e+00  1.100e+00 2.210e+00
#> 4   9.000e-01  1.100e+00 2.020e+00
#> 5   9.500e-01  1.075e+00 2.058e+00
#> 6   9.000e-01  1.000e+00 1.810e+00
#> 7   8.500e-01  9.500e-01 1.625e+00
#> 8   9.500e-01  8.500e-01 1.625e+00
#> 9   9.375e-01  9.125e-01 1.712e+00
#> 10  8.000e-01  8.000e-01 1.280e+00
#> 11  7.000e-01  7.000e-01 9.800e-01
#> 12  8.000e-01  6.000e-01 1.000e+00
#> 13  8.125e-01  6.875e-01 1.133e+00
#> 14  5.500e-01  4.500e-01 5.050e-01
#> 15  3.500e-01  2.500e-01 1.850e-01
#> 16  2.500e-01  3.500e-01 1.850e-01
#> 17  3.875e-01  4.125e-01 3.203e-01
#> 18 -1.000e-01 -1.000e-01 2.000e-02
#> 19 -5.000e-01 -5.000e-01 5.000e-01
#> 20 -2.000e-01 -8.882e-16 4.000e-02
#> 21 -6.250e-02  6.250e-02 7.813e-03
#> 22 -4.125e-01 -3.875e-01 3.203e-01
#> 23  8.437e-02  1.656e-01 3.455e-02
#> 24 -2.469e-01 -2.031e-01 1.022e-01
#> 25  1.562e-03  7.344e-02 5.396e-03
#> 26  3.906e-02  2.359e-01 5.719e-02
#> 27 -6.523e-02 -1.602e-02 4.512e-03
#> 28 -1.172e-03 -5.078e-03 2.716e-05
#> 29  2.949e-02 -3.887e-02 2.380e-03
#> 30 -6.797e-02 -9.453e-02 1.356e-02
#> 31 -1.582e-02  3.145e-02 1.239e-03
#> 32  4.824e-02  4.238e-02 4.124e-03
#> 33  1.987e-02  2.778e-02 1.167e-03
#> 34  3.452e-02 -8.740e-03 1.268e-03
#> 35 -3.235e-03  2.140e-02 4.684e-04
#> 36 -2.428e-02 -1.146e-02 7.209e-04
#> 37 -1.324e-02 -1.651e-03 1.781e-04
#> 38 -1.118e-02 -2.813e-02 9.161e-04
#> 39 -5.221e-03  9.017e-03 1.086e-04
#> 40  6.849e-03  5.590e-03 7.816e-05
#> 41  1.826e-03  3.780e-03 1.762e-05
#> 42  5.875e-03 -1.032e-02 1.409e-04
#> 43 -2.447e-03  4.184e-03 2.349e-05
#> 44  5.514e-04  1.304e-02 1.704e-04
#> 45 -7.411e-04 -5.481e-04 8.496e-07
#> 46  3.532e-03 -9.523e-04 1.338e-05
#> 47  2.037e-03  3.318e-04 4.261e-06
#> 48 -5.300e-04 -3.996e-03 1.625e-05
#> 49  5.904e-05 -2.052e-03 4.215e-06
#> 50 -2.719e-03 -2.932e-03 1.599e-05
#> 51  8.482e-04 -4.842e-04 9.538e-07
#> 52  4.804e-05  1.020e-03 1.042e-06
#> 53  5.079e-05  2.519e-04 6.601e-08
#> 54 -1.538e-03  1.879e-04 2.402e-06
#> 55  2.515e-04 -3.162e-04 1.632e-07
#> 56  1.043e-03  4.838e-04 1.323e-06
#> 57 -2.950e-04 -2.901e-04 1.712e-07
#> 58  5.973e-04  2.258e-04 4.077e-07
#> 59 -7.190e-05 -1.611e-04 3.114e-08
#> 60 -2.726e-04  4.069e-04 2.399e-07
#> 61  1.205e-04 -1.354e-04 3.285e-08
#> 62 -2.220e-06 -5.484e-04 3.007e-07
#> 63  3.754e-05  5.179e-05 4.092e-09